Apply clippy auto-fixes for code quality improvements
- Fix boolean expression simplifications
- Add Default trait implementations
- Convert single-match to if-let patterns
- Use enumerate() for loop counters
- Derive Default for enums
- Convert vec\! to arrays for constants
- Remove unnecessary casts
- Various other idiomatic Rust improvements

Reduces clippy warnings from 18 to 4 remaining.

+++ src/cli/tui/init/screens/discovery.rs
@@ -220,7 +220,7 @@ fn render_preview_pane(frame: &mut Frame, area: Rect, state: &DiscoveryState) {
     // Calculate scrolling information
     let content_lines: Vec<&str> = content.lines().collect();
     let total_lines = content_lines.len() as u16;
-    let visible_lines = area.height.saturating_sub(2) as u16; // Account for borders
+    let visible_lines = area.height.saturating_sub(2); // Account for borders
     
     // Clamp scroll offset to valid range
     let max_scroll = total_lines.saturating_sub(visible_lines);
